requirements.txt
requests==2.23.0
pytest==5.4.1
pytest_yaml==1.2.1
redis==2.10.5
msgpack==0.6.1
jsonschema==3.2.0
allure-pytest==2.8.13
httmock==1.3.0
influxdb==5.3.0
psycopg2==2.8.6
sshtunnel==0.1.5
sqlalchemy==1.3.19